pyproject.toml
[project]
name = "speakpy"
version = "0.1.0"
description = "TTS Server Daemon"
readme = "README.md"
authors = [
    { name = "Toprak", email = "efe.akkilic@ozu.edu.tr" }
]
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.36.0",
    "pedalboard>=0.9.19",
    "piper-tts>=1.3.0",
    "platformdirs>=4.5.0",
    "sounddevice>=0.5.3",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'"
]

[tool.uv.build-backend]
module-name = "speakpy"
